chore($cascader): function extraction and fix error
